requirements.txt
Pillow==10.0.1
torch==1.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.66.1
transformers==4.37.2
